+++ doc/3d-snapshot/GetFieldPtr.c
@@ -22,7 +22,7 @@ GetFieldPtrValidateT(
 )
 {
   /* Validating field f1 */
-  BOOLEAN hasEnoughBytes0 = (uint64_t)(uint32_t)(uint8_t)10U <= (InputLength - StartPosition);
+  BOOLEAN hasEnoughBytes0 = (uint64_t)(uint32_t)10U <= (InputLength - StartPosition);
   uint64_t positionAfterT;
   if (!hasEnoughBytes0)
   {
@@ -33,24 +33,24 @@ GetFieldPtrValidateT(
   else
   {
     uint8_t *truncatedInput = Input;
-    uint64_t truncatedInputLength = StartPosition + (uint64_t)(uint32_t)(uint8_t)10U;
+    uint64_t truncatedInputLength = StartPosition + (uint64_t)(uint32_t)10U;
     uint64_t result = StartPosition;
     while (TRUE)
     {
       uint64_t position = result;
       BOOLEAN ite;
-      if (!((uint64_t)1U <= (truncatedInputLength - position)))
+      if (!(1ULL <= (truncatedInputLength - position)))
       {
         ite = TRUE;
       }
       else
       {
         /* Checking that we have enough space for a UINT8, i.e., 1 byte */
-        BOOLEAN hasBytes = (uint64_t)1U <= (truncatedInputLength - position);
+        BOOLEAN hasBytes = 1ULL <= (truncatedInputLength - position);
         uint64_t positionAfterT;
         if (hasBytes)
         {
-          positionAfterT = position + (uint64_t)1U;
+          positionAfterT = position + 1ULL;
         }
         else
         {
@@ -107,7 +107,7 @@ GetFieldPtrValidateT(
     return positionAfterf1;
   }
   /* Validating field f2 */
-  BOOLEAN hasEnoughBytes = (uint64_t)(uint32_t)(uint8_t)20U <= (InputLength - positionAfterf1);
+  BOOLEAN hasEnoughBytes = (uint64_t)(uint32_t)20U <= (InputLength - positionAfterf1);
   uint64_t positionAfterT0;
   if (!hasEnoughBytes)
   {
@@ -118,24 +118,24 @@ GetFieldPtrValidateT(
   else
   {
     uint8_t *truncatedInput = Input;
-    uint64_t truncatedInputLength = positionAfterf1 + (uint64_t)(uint32_t)(uint8_t)20U;
+    uint64_t truncatedInputLength = positionAfterf1 + (uint64_t)(uint32_t)20U;
     uint64_t result = positionAfterf1;
     while (TRUE)
     {
       uint64_t position = result;
       BOOLEAN ite;
-      if (!((uint64_t)1U <= (truncatedInputLength - position)))
+      if (!(1ULL <= (truncatedInputLength - position)))
       {
         ite = TRUE;
       }
       else
       {
         /* Checking that we have enough space for a UINT8, i.e., 1 byte */
-        BOOLEAN hasBytes = (uint64_t)1U <= (truncatedInputLength - position);
+        BOOLEAN hasBytes = 1ULL <= (truncatedInputLength - position);
         uint64_t positionAfterT;
         if (hasBytes)
         {
-          positionAfterT = position + (uint64_t)1U;
+          positionAfterT = position + 1ULL;
         }
         else
         {
