+++ doc/3d-snapshot/Smoker.c
@@ -21,11 +21,11 @@ SmokerValidateSmoker(
 )
 {
   /* Checking that we have enough space for a UINT32, i.e., 4 bytes */
-  BOOLEAN hasBytes0 = (uint64_t)4U <= (InputLength - StartPosition);
+  BOOLEAN hasBytes0 = 4ULL <= (InputLength - StartPosition);
   uint64_t positionAfternone;
   if (hasBytes0)
   {
-    positionAfternone = StartPosition + (uint64_t)4U;
+    positionAfternone = StartPosition + 4ULL;
   }
   else
   {
@@ -41,7 +41,7 @@ SmokerValidateSmoker(
   else
   {
     uint32_t none = Load32Le(Input + (uint32_t)StartPosition);
-    BOOLEAN noneConstraintIsOk = none >= (uint32_t)(uint8_t)21U;
+    BOOLEAN noneConstraintIsOk = none >= (uint32_t)21U;
     uint64_t
     positionAfternone1 = EverParseCheckConstraintOk(noneConstraintIsOk, positionAfternone);
     if (EverParseIsError(positionAfternone1))
@@ -52,11 +52,11 @@ SmokerValidateSmoker(
     {
       /* Validating field cigarettesConsumed */
       /* Checking that we have enough space for a UINT8, i.e., 1 byte */
-      BOOLEAN hasBytes = (uint64_t)1U <= (InputLength - positionAfternone1);
+      BOOLEAN hasBytes = 1ULL <= (InputLength - positionAfternone1);
       uint64_t positionAfterSmoker0;
       if (hasBytes)
       {
-        positionAfterSmoker0 = positionAfternone1 + (uint64_t)1U;
+        positionAfterSmoker0 = positionAfternone1 + 1ULL;
       }
       else
       {
